feat: split nodes into separate partitions
